requirements.txt
opencv-python
numpy
pillow
torch
transformers
ultralytics
pytest
